+++ src/test/java/no/fintlabs/consumer/links/LinkServiceTest.java
@@ -1,7 +1,7 @@
 package no.fintlabs.consumer.links;
 
 import no.fint.model.felles.kompleksedatatyper.Identifikator;
-import no.fint.model.resource.FintResources;
+import no.fintlabs.model.resource.FintResources;
 import no.fint.model.resource.Link;
 import no.fint.model.resource.utdanning.elev.ElevResource;
 import no.fintlabs.consumer.links.nested.NestedLinkService;
